src/main/java/ru/elcoder/leetcode/MaximumDepthOfBinaryTreeSolution.java
package ru.elcoder.leetcode;

import ru.elcoder.leetcode.models.DifficultyLevel;
import ru.elcoder.leetcode.models.Leetcode;
import ru.elcoder.leetcode.models.TreeNode;

@Leetcode(
        difficulty = DifficultyLevel.Easy,
        number = 104,
        title = "Maximum Depth of Binary Tree",
        url = "https://leetcode.com/problems/maximum-depth-of-binary-tree/"
)
public class MaximumDepthOfBinaryTreeSolution {
    public int maxDepth(TreeNode root) {
        return root == null ? 0 : (1 + Math.max(maxDepth(root.left), maxDepth(root.right)));
    }
}
